backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
sqlalchemy
python-dotenv
psycopg2-binary
pyarrow
pandas
geopandas